+++ src/main/kotlin/io/github/barsia/speqa/registry/IdType.kt
@@ -1,0 +1,6 @@
+package io.github.barsia.speqa.registry
+
+enum class IdType {
+    TEST_CASE,
+    TEST_RUN,
+}